rename cellfinder workflow to cellfinder_core (to separate from cellfinder-brainreg CLI while keeping backwards comp) WIP
