+++ tests/cellfinder_core/test_integration/brainglobe_benchmarks/test_cellfinder.py
