+++ tests/cellfinder_core/test_integration/brainglobe_workflows/test_cellfinder.py
@@ -5,7 +5,7 @@ from typing import Optional
 
 import pytest
 
-from brainglobe_workflows.cellfinder import main
+from brainglobe_workflows.cellfinder_core.cellfinder_core import main
 
 
 @pytest.mark.parametrize(
