+++ tests/cellfinder_core/test_unit/brainglobe_workflows/conftest.py
@@ -10,6 +10,8 @@ def custom_logger_name() -> str:
     str
         Name of custom logger
     """
-    from brainglobe_workflows.utils import __name__ as logger_name
+    from brainglobe_workflows.cellfinder_core.utils import (
+        __name__ as logger_name,
+    )
 
     return logger_name
